[feature] able to add user router

+++ db/sqlc/querier.go
@@ -10,8 +10,8 @@ import (
 
 type Querier interface {
 	CreateUsers(ctx context.Context, arg CreateUsersParams) (User, error)
-	DeleteUsers(ctx context.Context, id int32) error
-	GetUserById(ctx context.Context, id int32) (User, error)
+	DeleteUsers(ctx context.Context, id int64) error
+	GetUserById(ctx context.Context, id int64) (User, error)
 	GetUserByUserName(ctx context.Context, username string) (User, error)
 	ListUsers(ctx context.Context, arg ListUsersParams) ([]User, error)
 	UpdateUsers(ctx context.Context, arg UpdateUsersParams) (User, error)
